refactored out tests to unit tests
